Splitted v compiler into separated files
